+++ components/global-components/Footer.js
@@ -2,7 +2,7 @@ import React from "react";
 import Link from "next/link";
 
 const Footer = () => {
-  const version = "v2.5.2";
+  const version = "v2.5.3";
 
   return (
     <footer className="text-center w-full ">